requirements.txt
Flask
supertokens-python
Flask-Cors
python-dotenv
SQLAlchemy==1.4.41
psycopg2-binary==2.9.3
validator.py==1.3.0
Flask-HTTPAuth
pytest
Faker==15.1.1